# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateMusicVideoRequest:
    properties:
      artist:
        type: string
      artist_id:
        type: integer
      description:
        type: string
      duration:
        description: 'Format: MM:SS atau HH:MM:SS'
        type: string
      genre:
        type: string
      release_date:
        description: 'Format: "2006-01-02"'
        type: string
      thumbnail:
        type: string
      title:
        type: string
      video_url:
        type: string
    required:
    - artist
    - artist_id
    - duration
    - genre
    - release_date
    - title
    - video_url
    type: object
...
  handlers.UpdateMusicVideoRequest:
    properties:
      artist:
        type: string
      artist_id:
        type: integer
      description:
        type: string
      duration:
        description: 'Format: MM:SS atau HH:MM:SS'
        type: string
      genre:
        type: string
      release_date:
        description: 'Format: "2006-01-02"'
        type: string
      thumbnail:
        type: string
      title:
        type: string
      video_url:
        type: string
    type: object
...
info:
  contact:
    email: support@soundcave.com
    name: API Support
  description: API documentation for SoundCave Backend - Music streaming platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SoundCave Backend API
  version: "1.0"
paths:
...
  /music-videos:
    get:
      consumes:
      - application/json
      description: Get paginated list of music videos with filtering and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by artist ID
        in: query
        name: artist_id
        type: integer
      - description: Filter by genre
        in: query
        name: genre
        type: string
      - description: Search by title or artist
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort field
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all music videos
      tags:
      - MusicVideos
    post:
      consumes:
      - application/json
      description: Create a new music video
      parameters:
      - description: Music Video Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMusicVideoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new music video
      tags:
      - MusicVideos
  /music-videos/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a music video by ID
      parameters:
      - description: Music Video ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete music video
      tags:
      - MusicVideos
    get:
      consumes:
      - application/json
      description: Get music video details by ID
      parameters:
      - description: Music Video ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get music video by ID
      tags:
      - MusicVideos
    put:
      consumes:
      - application/json
      description: Update music video information
      parameters:
      - description: Music Video ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Music Video Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateMusicVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update music video
      tags:
      - MusicVideos
...
securityDefinitions:
  BearerAuth:
    description: 'To authenticate, provide the API token with the ''Bearer '' prefix.
      Example: ''Bearer your_token_here'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"